requirements.txt
fastapi
uvicorn
transformers
newspaper3k
beautifulsoup4
pydantic